requirements.txt
colorama==0.4.6
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0

# Django & REST Framework
django>=6.0.1
djangorestframework>=3.16.1

# Database
asgiref>=3.9.1
sqlparse>=0.5.0
